+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 ## Unreleased
 
+### Added
+
+* `Carbonite.Migration.drop_tables/1` allows to drop the carbonite audit trail without removing the schema
+
 ### Fixed
 
 * Fixed ignore mode when override_transaction_id is NULL
+
+### Changed
+
+* Renamed the option that can be passed to `Carbonite.Migration.drop_schema/1` from `prefix` to `carbonite_prefix`
 
 ## [0.2.1] - 2021-10-10